idk what i wrote

+++ lib/screen/home_course_list.dart
@@ -8,12 +8,14 @@ import 'package:aula/screen/chat.dart';
 import 'package:aula/screen/course_screen.dart';
 import 'package:aula/screen/notif.dart';
 import 'package:aula/screen/profile_screen.dart';
+// import 'package:downloads_path_provider/downloads_path_provider.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:aula/bloc/cardlist/cardlist_bloc.dart';
 import 'package:image_picker/image_picker.dart';
+import 'package:permission_handler/permission_handler.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:path/path.dart';
@@ -69,7 +71,7 @@ class _CourseListHomeState extends State<CourseListHome> {
             Positioned(top: 0, child: UserDetailsCard()),
             Positioned.directional(
               textDirection: TextDirection.ltr,
-              top: 220,
+              top: 224,
               child: Container(
                 padding: EdgeInsets.zero,
                 width: size.width,
@@ -116,15 +118,15 @@ class MenuAppBar extends StatefulWidget {
 }
 
 class _MenuAppBarState extends State<MenuAppBar> {
-  final color = Colors.white;
+  final iconColor = Colors.white;
   double top = 0.0, opacity = 0.0;
   @override
   void initState() {
     widget.pc.addListener(() {
       var off = widget.pc.offset;
-      if (off <= 94) {
+      if (off <= 88) {
         setState(() {
-          opacity = off != 0.0 ? off / 94 : 0.0;
+          opacity = off != 0.0 ? off / 88 : 0.0;
         });
       } else {
         opacity = 1.0;
@@ -144,9 +146,9 @@ class _MenuAppBarState extends State<MenuAppBar> {
     return Positioned(
       top: top,
       child: Container(
-        padding: EdgeInsets.fromLTRB(24, 24, 14, 0),
+        padding: EdgeInsets.fromLTRB(24, 24, 14, 8),
         width: size.width,
-        color: Colors.red[900].withOpacity(opacity),
+        color: Colors.purple[600].withOpacity(opacity),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: <Widget>[
@@ -154,7 +156,7 @@ class _MenuAppBarState extends State<MenuAppBar> {
               onTap: widget.callback,
               child: Container(
                 padding: EdgeInsets.all(4),
-                child: Icon(Icons.menu, color: color),
+                child: Icon(Icons.menu, color: iconColor),
               ),
             ),
             Row(children: [
@@ -169,7 +171,7 @@ class _MenuAppBarState extends State<MenuAppBar> {
                       },
                       icon: Icon(
                         Icons.notifications,
-                        color: color,
+                        color: iconColor,
                       )),
                   Positioned(
                       top: 8,
@@ -195,7 +197,7 @@ class _MenuAppBarState extends State<MenuAppBar> {
                       },
                       icon: Icon(
                         Icons.chat_bubble,
-                        color: color,
+                        color: iconColor,
                       )),
                   Positioned(
                       top: 8,
@@ -224,10 +226,16 @@ class UserDetailsCard extends StatelessWidget {
         child: Container(
           decoration: BoxDecoration(
               color: Colors.purple,
+              boxShadow: [
+                BoxShadow(
+                  spreadRadius: 0.0,
+                  blurRadius: 8.0,
+                )
+              ],
               borderRadius: BorderRadius.only(
                   bottomLeft: Radius.circular(48),
                   bottomRight: Radius.circular(48))),
-          padding: EdgeInsets.fromLTRB(24.0, 64.0, 24.0, 16.0),
+          padding: EdgeInsets.fromLTRB(24.0, 68.0, 24.0, 8.0),
           width: width,
           height: 160,
           child: Row(
@@ -305,24 +313,38 @@ class _ProfilePictState extends State<ProfilePict> {
   Future setup() async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
     var dir = prefs.getString('profile_picture');
+    print(dir);
     return _image = File(dir);
   }
 
   Future getImage() async {
+    var str = await Permission.mediaLibrary.request();
     var pickedFile = await picker.getImage(source: ImageSource.gallery);
     SharedPreferences prefs = await SharedPreferences.getInstance();
     if (pickedFile == null) return;
     //get sys dir
-    Directory appDocDir = await getApplicationDocumentsDirectory();
+    var listAppDocDir =
+        await getExternalStorageDirectories(type: StorageDirectory.pictures);
+    Directory appDocDir = listAppDocDir[0];
+
+    Directory appDir = Directory('/storage/emulated/0/Downloads/');
+
     String appDocPath = appDocDir.path;
+    // '/storage/emulated/0/Pictures';
+    print(Permission.values);
+    print(appDir.existsSync());
 
 // copy the file to a new path
     var extensi = (extension(pickedFile.path));
     try {
       var tgl = DateTime.now();
       String fileLama = prefs.getString('profile_picture');
+      print('$fileLama');
       if (fileLama != null) {
-        await File(fileLama).delete();
+        if (await File(fileLama).exists()) {
+          await File(fileLama).delete();
+        }
+        prefs.remove('profile_picture');
       }
       String namaFile = tgl.minute.toString() + tgl.second.toString();
       String newDir = '$appDocPath/$namaFile' 'pp$extensi';
@@ -439,12 +461,12 @@ class CourseCard extends StatelessWidget {
                 margin: EdgeInsets.symmetric(vertical: 6, horizontal: 18),
                 decoration: BoxDecoration(
                   color: Colors.white,
-                  // boxShadow: [
-                  //   BoxShadow(
-                  //       color: Colors.black12,
-                  //       blurRadius: 2.0,
-                  //       spreadRadius: 2.0),
-                  // ],
+                  boxShadow: [
+                    BoxShadow(
+                        color: Colors.black12,
+                        blurRadius: 2.0,
+                        spreadRadius: 2.0),
+                  ],
                   borderRadius: BorderRadius.circular(16),
                 ),
                 child: Column(
@@ -581,19 +603,20 @@ class ViewControl extends StatefulWidget {
 }
 
 class _ViewControlState extends State<ViewControl> {
-  var viewControlloffset = 166.0;
+  var viewControlloffset = 172.0;
   int filterVal = 0;
   int sortByVal = 0;
+  var style = TextStyle(fontSize: 12.0);
   @override
   void initState() {
     this.widget.scont.addListener(() {
-      if (this.widget.scont.offset >= 90) {
+      if (this.widget.scont.offset >= 84) {
         setState(() {
-          viewControlloffset = this.widget.scont.offset + 68;
+          viewControlloffset = this.widget.scont.offset + 78;
         });
-      } else if (viewControlloffset != 165.0) {
+      } else if (viewControlloffset != 170.0) {
         setState(() {
-          viewControlloffset = 165.0;
+          viewControlloffset = 170.0;
         });
       }
     });
@@ -606,57 +629,95 @@ class _ViewControlState extends State<ViewControl> {
     return Positioned(
       top: viewControlloffset,
       child: Container(
-        padding: EdgeInsets.symmetric(horizontal: 14),
+        padding: EdgeInsets.symmetric(horizontal: 10),
         decoration: BoxDecoration(
-          boxShadow: [BoxShadow(offset: Offset(0, 1), blurRadius: 2.0)],
+          boxShadow: [
+            BoxShadow(
+                offset: Offset(0, 4), blurRadius: 4.0, color: Colors.grey),
+          ],
           color: Colors.white,
         ),
         width: size.width,
-        child: SingleChildScrollView(
-          scrollDirection: Axis.horizontal,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              Row(
-                children: <Widget>[
-                  Icon(Icons.filter_list),
-                  DropdownButton(
-                      value: filterVal,
-                      items: [
-                        DropdownMenuItem(
-                            child: Text('All(Except Hidden)'), value: 0),
-                        DropdownMenuItem(child: Text('In progress'), value: 1),
-                        DropdownMenuItem(child: Text('Future'), value: 2),
-                        DropdownMenuItem(child: Text('Past'), value: 3),
-                        DropdownMenuItem(child: Text('Starred'), value: 4),
-                        DropdownMenuItem(child: Text('Hidden'), value: 5),
-                      ],
-                      onChanged: (value) {
-                        setState(() {
-                          filterVal = value;
-                        });
-                      }),
-                ],
-              ),
-              Row(
-                children: <Widget>[
-                  Icon(Icons.sort),
-                  DropdownButton(
-                      value: sortByVal,
-                      items: [
-                        DropdownMenuItem(child: Text('Course Name'), value: 0),
-                        DropdownMenuItem(
-                            child: Text('Last Accessed'), value: 1),
-                      ],
-                      onChanged: (value) {
-                        setState(() {
-                          sortByVal = value;
-                        });
-                      }),
-                ],
-              ),
-            ],
-          ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: <Widget>[
+            Row(
+              children: <Widget>[
+                Icon(Icons.filter_list),
+                DropdownButton(
+                    value: filterVal,
+                    items: [
+                      DropdownMenuItem(
+                          child: Text(
+                            'All(Except Hidden)',
+                            style: style,
+                          ),
+                          value: 0),
+                      DropdownMenuItem(
+                          child: Text(
+                            'In progress',
+                            style: style,
+                          ),
+                          value: 1),
+                      DropdownMenuItem(
+                          child: Text(
+                            'Future',
+                            style: style,
+                          ),
+                          value: 2),
+                      DropdownMenuItem(
+                          child: Text(
+                            'Past',
+                            style: style,
+                          ),
+                          value: 3),
+                      DropdownMenuItem(
+                          child: Text(
+                            'Starred',
+                            style: style,
+                          ),
+                          value: 4),
+                      DropdownMenuItem(
+                          child: Text(
+                            'Hidden',
+                            style: style,
+                          ),
+                          value: 5),
+                    ],
+                    onChanged: (value) {
+                      setState(() {
+                        filterVal = value;
+                      });
+                    }),
+              ],
+            ),
+            Row(
+              children: <Widget>[
+                Icon(Icons.sort),
+                DropdownButton(
+                    value: sortByVal,
+                    items: [
+                      DropdownMenuItem(
+                          child: Text(
+                            'Course Name',
+                            style: style,
+                          ),
+                          value: 0),
+                      DropdownMenuItem(
+                          child: Text(
+                            'Last Accessed',
+                            style: style,
+                          ),
+                          value: 1),
+                    ],
+                    onChanged: (value) {
+                      setState(() {
+                        sortByVal = value;
+                      });
+                    }),
+              ],
+            ),
+          ],
         ),
       ),
     );
